refactor: Move pasta 'Produtos' para dentro da pasta 'Evento' e atualiza rotas e path dos icones
